More html responses to template, use POST
