+++ pyproject.toml
@@ -13,6 +13,7 @@ ipython = "^7.6.1"
 numpy = "^1.16.4"
 sacn = "^1.3"
 ola = "^0.10.7"
+jinja2 = "^2.10"
 
 [tool.poetry.dev-dependencies]
 pytest = "^3.0"
