+++ web/templates/index.html
@@ -1,0 +1,37 @@
+<!doctype html>
+<html lang="en">
+<head>
+  <meta charset="utf-8">
+  <title>Show Control</title>
+</head>
+<body>
+
+<h1>Show Control</h1>
+
+<p>Status: {{status}}</p>
+
+<h3>Choose a show</h3>
+<ul>
+{% for show in shows %}
+  <li><form action="run_show" method="post">
+    <button name="show_name" value="{{show}}">{{show}}</button>
+  </form></li>
+{% endfor %}
+</ul>
+
+<br/>
+
+<form action="clear_show" method="post">
+  <button>Clear Show</button>
+</form>
+
+<br/>
+
+<h3>Change Show Cycle Time (seconds)</h3>
+<form action="change_run_time" method="post">
+  Seconds: <input id=run_time type=number name=run_time value=60>
+  <input type=submit>
+</form>
+
+</body>
+</html>
